fix memberships pages wehere the spaceLevel was still expected as number (#7071)
